Prevent the compatibility mode in Oculus Quest
The latest firmware update introduced some kind of compatibility mode
for (probably?) Oculus Go which activates if you don't specify the
android.hardware.vr.headtracking feature. This broke most of sideloaded
apps as this was usually only truly required when submitting the app to
Oculus Store.

Note that I specified required="false" to make sure we still can run on
3DoF headsets. Most people are talking about using the line with
required="true", but this is the correct way for hybrid apps according
to Oculus docs (and it works)

Credits to everybody who figured out this issue before me